+++ core/services/ocr2/plugins/ccip/internal/ccipdata/factory/onramp.go
@@ -12,7 +12,6 @@ import (
 	ccipconfig "github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/config"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipcalc"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata"
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_0_0"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_2_0"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_5_0"
 )
@@ -44,15 +43,6 @@ func initOrCloseOnRampReader(lggr logger.Logger, versionFinder VersionFinder, so
 	lggr.Infof("Initializing onRamp for version %v", version.String())
 
 	switch version.String() {
-	case ccipdata.V1_0_0:
-		onRamp, err := v1_0_0.NewOnRamp(lggr, sourceSelector, destSelector, onRampAddrEvm, sourceLP, source)
-		if err != nil {
-			return nil, err
-		}
-		if closeReader {
-			return nil, onRamp.Close()
-		}
-		return onRamp, onRamp.RegisterFilters()
 	case ccipdata.V1_2_0:
 		onRamp, err := v1_2_0.NewOnRamp(lggr, sourceSelector, destSelector, onRampAddrEvm, sourceLP, source)
 		if err != nil {
