+++ core/services/ocr2/plugins/ccip/pkg/leafer/leafer.go
@@ -10,9 +10,7 @@ import (
 	"github.com/smartcontractkit/chainlink-common/pkg/hashutil"
 
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp"
-	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp_1_0_0"
 	"github.com/smartcontractkit/chainlink/v2/core/gethwrappers/ccip/generated/evm_2_evm_onramp_1_2_0"
-	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_0_0"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_2_0"
 	"github.com/smartcontractkit/chainlink/v2/core/services/ocr2/plugins/ccip/internal/ccipdata/v1_5_0"
 )
@@ -26,7 +24,6 @@ type LeafHasher interface {
 type Version string
 
 const (
-	V1_0_0 Version = "v1_0_0"
 	V1_2_0 Version = "v1_2_0"
 	V1_5_0 Version = "v1_5_0"
 )
@@ -34,13 +31,6 @@ const (
 // MakeLeafHasher is a factory function to construct the onramp implementing the HashLeaf function for a given version.
 func MakeLeafHasher(ver Version, cl bind.ContractBackend, sourceChainSelector uint64, destChainSelector uint64, onRampId common.Address, ctx hashutil.Hasher[[32]byte]) (LeafHasher, error) {
 	switch ver {
-	case V1_0_0:
-		or, err := evm_2_evm_onramp_1_0_0.NewEVM2EVMOnRamp(onRampId, cl)
-		if err != nil {
-			return nil, err
-		}
-		h := v1_0_0.NewLeafHasher(sourceChainSelector, destChainSelector, onRampId, ctx, or)
-		return h, nil
 	case V1_2_0:
 		or, err := evm_2_evm_onramp_1_2_0.NewEVM2EVMOnRamp(onRampId, cl)
 		if err != nil {
